Add path and date parameters to news_to_html_converter and news_to_pdf_converter functions, add removing extra quotes from links and images links of description
To work with user passed path for file saving and to
make it possible to filter news on date path and date
parameters are added to news_to_html_converter and
news_to_pdf_converter functions. When getting the cached
item description links extra quotes were passed from Panda
DataFrame and as the result some links in pdf were not
working: so some code to get rid of them is added.

Version 2.0.5